requirements.txt
oauthlib==1.0.0
Pillow==2.9.0
requests==2.7.0
requests-oauthlib==0.5.0
tendo==0.2.7
twitter==1.17.0
TwitterAPI==2.3.3.1